Move topology HTML assets out of renderers
